LPS-124147 Forms responsibility components moved to the form-web module
Despite having many files being moved and edited, there are few changes the vast majority in the import and relocation of files to the respective modules.

The components that are exclusive to the Forms product are inside the form-web, components of `custom/forms` are components that may have become part of the Data Engine core in the future but the only customer is the product with the respective folder name or the component is too coupled to the product to be part of the core at the moment.
