+++ modules/apps/dynamic-data-mapping/dynamic-data-mapping-form-renderer/package.json
@@ -1,6 +1,14 @@
 {
 	"dependencies": {
+		"@clayui/button": "3.6.0",
+		"@clayui/data-provider": "3.3.10",
+		"@clayui/drop-down": "3.25.0",
+		"@clayui/icon": "3.1.0",
 		"@clayui/layout": "3.3.0",
+		"@clayui/loading-indicator": "3.2.0",
+		"@clayui/pagination": "3.25.0",
+		"@clayui/tabs": "3.3.5",
+		"@clayui/tooltip": "3.4.5",
 		"classnames": "2.2.6",
 		"dom-align": "1.10.4",
 		"frontend-js-web": "*",
@@ -8,8 +16,8 @@
 		"metal-component": "2.16.8",
 		"metal-soy": "2.16.9",
 		"react": "16.12.0",
-		"react-router-dom": "^5.0.1",
-		"object-hash": "^1.3.0"
+		"react-dnd": "11.1.1",
+		"react-dnd-html5-backend": "11.1.1"
 	},
 	"jest": {
 		"modulePathIgnorePatterns": [
@@ -19,7 +27,7 @@
 			"metal-jest-serializer"
 		]
 	},
-	"main": "index.js",
+	"main": "js/index.js",
 	"name": "dynamic-data-mapping-form-renderer",
 	"private": true,
 	"scripts": {
