+++ modules/apps/dynamic-data-mapping/dynamic-data-mapping-form-renderer/src/main/resources/META-INF/resources/js/core/config/initialConfigState.es.js
@@ -31,13 +31,9 @@ const getFieldNameGenerator = (pages, generateFieldNameUsingFieldLabel) => (
 		generateFieldNameUsingFieldLabel
 	);
 
-export const COMMON_INITIAL_CONFIG_STATE = {
+export const INITIAL_CONFIG_STATE = {
 	cache: {},
 	defaultLanguageId: themeDisplay.getLanguageId(),
-};
-
-export const BUILDER_INITIAL_CONFIG_STATE = {
-	...COMMON_INITIAL_CONFIG_STATE,
 	generateFieldNameUsingFieldLabel: false,
 	getFieldNameGenerator,
 };
