+++ modules/apps/dynamic-data-mapping/dynamic-data-mapping-form-renderer/src/main/resources/META-INF/resources/js/core/reducers/index.es.js
@@ -12,23 +12,9 @@
  * details.
  */
 
-const PAGINATION = {
-	CHANGE: 'pagination_change',
-	NEXT: 'pagination_next',
-	PREVIOUS: 'pagination_previous',
-};
-
-const PAGE = {
-	ADD: 'page_add',
-	DELETE: 'page_delete',
-	DESCRIPTION_CHANGE: 'page_description_change',
-	RESET: 'page_reset',
-	SWAP: 'page_swap',
-	TITLE_CHANGE: 'page_title_change',
-};
-
-export const EVENT_TYPES = {
-	PAGE,
-	PAGINATION,
-	SUCCESS_PAGE: 'success_page',
-};
+export {default as activePageReducer} from './activePageReducer.es';
+export {default as dragAndDropReducer} from './dragAndDropReducer.es';
+export {default as fieldEditableReducer} from './fieldEditableReducer.es';
+export {default as fieldReducer} from './fieldReducer.es';
+export {default as languageReducer} from './languageReducer.es';
+export {default as pagesStructureReducer} from './pagesStructureReducer.es';
