+++ modules/apps/dynamic-data-mapping/dynamic-data-mapping-form-renderer/src/main/resources/META-INF/resources/js/custom/form/reducers/index.es.js
@@ -12,10 +12,6 @@
  * details.
  */
 
-export {default as elementSetReducer} from './elementSetReducer.es';
 export {default as pageReducer} from './pageReducer.es';
 export {default as pageValidationReducer} from './pageValidationReducer.es';
 export {default as paginationReducer} from './paginationReducer.es';
-export {default as rulesReducer} from './rulesReducer.es';
-export {default as sidebarReducer} from './sidebarReducer.es';
-export {default as formInfoReducer} from './formInfoReducer.es';
