+++ modules/apps/dynamic-data-mapping/dynamic-data-mapping-form-web/package.json
@@ -1,4 +1,18 @@
 {
+	"dependencies": {
+		"@clayui/alert": "3.5.1",
+		"@clayui/button": "3.6.0",
+		"@clayui/data-provider": "3.3.10",
+		"@clayui/form": "3.14.4",
+		"@clayui/layout": "3.3.0",
+		"@clayui/link": "3.2.0",
+		"@clayui/modal": "3.8.5",
+		"react": "16.12.0",
+		"react-dnd": "11.1.1",
+		"react-dnd-html5-backend": "11.1.1",
+		"react-router-dom": "^5.0.1",
+		"object-hash": "^1.3.0"
+	},
 	"jest": {
 		"modulePathIgnorePatterns": [
 			"/__mock__/"
